set tab heaeders
