+++ MongoMS/MongoMS.csproj
@@ -111,7 +111,9 @@
       <SubType>Code</SubType>
     </Compile>
     <Compile Include="AppBootstrapper.cs" />
+    <Compile Include="CloseTabbedViewAction.cs" />
     <Compile Include="Controls\LastTabSelectionTabControl.cs" />
+    <Compile Include="HeaderTextBlock.cs" />
     <Compile Include="Shell.xaml.cs">
       <DependentUpon>Shell.xaml</DependentUpon>
     </Compile>
@@ -153,6 +155,12 @@
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\MongoMS.Common\MongoMS.Common.csproj">
+      <Project>{1C05DC85-FE00-4BC1-A3C6-0D924F937362}</Project>
+      <Name>MongoMS.Common</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(SolutionDir)\.nuget\NuGet.targets" Condition="Exists('$(SolutionDir)\.nuget\NuGet.targets')" />
